server/requirements.txt
fastapi
uvicorn[standard]
python-multipart
websockets
nats-py
boto3
